версия SW-1g
sw_time_limit = 500 (было 450)

в sw_do-change добавлено
disable_e0();
disable_e1();

чистка лишнего мусора

+++ Marlin/Marlin.h
@@ -255,7 +255,7 @@ extern unsigned char fanSpeedSoftPwm;
   extern int sw_current; //значение тока
   extern int sw_current_max; // Максимальное значение тока, для отладки
   extern unsigned long sw_on_timer;
-  extern int sw_test;
+  extern unsigned long sw_test;
   extern int sw_timeout;
   extern int sw_calibrate_z_now;
   extern int reload_filament_sw;
